requirements.txt
google-adk
google-genai>=1.0.0
pandas
google-cloud-storage
google-cloud-secret-manager
requests
python-dotenv
uvicorn[standard]
fastapi
opentelemetry-instrumentation-google-genai
